# docs/mobile_v1_openapi.yaml
openapi: 3.0.3
info:
  title: FichaYa - Mobile API v1
  version: 1.25.0
  description: |
    API REST para la aplicacion Flutter de control de asistencia.
    Todos los endpoints (excepto auth) requieren header `Authorization: Bearer <token>`.
    Los modulos complementarios Feedback (`/api/v1/feedback`) y SKAP (`/api/skap`)
    usan el mismo Bearer JWT mobile, aunque no cuelgan del prefijo `/api/v1/mobile`.
    La campana de Alertas del home agrupa adelantos, pedidos de mercaderia y feedback
    para mostrar novedades o aprobaciones pendientes sin cambiar los flujos existentes.
  contact:
    name: Equipo Backend
servers:
  - url: http://localhost:5000/api/v1/mobile
    description: Local
  - url: https://control-asistencia-backend-8gle.onrender.com/api/v1/mobile
    description: Produccion

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OkErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string

    # --- Adelanto ---
...
    PedidoMercaderiaLinea:
      type: object
      properties:
        id:
          type: integer
          nullable: true
        articulo_id:
          type: integer
        codigo_articulo:
          type: string
        descripcion:
          type: string
        unidades_por_bulto:
          type: integer
        cantidad_bultos:
          type: integer
          minimum: 0
        cantidad_unidades:
          type: integer
          minimum: 0
          description: Unidades sueltas adicionales a los bultos completos.
        total_unidades:
          type: integer
          minimum: 0
          description: cantidad_bultos * unidades_por_bulto + cantidad_unidades.
    PedidoMercaderiaItem:
      type: object
      properties:
        id:
          type: integer
        periodo:
          type: string
          example: "2026-04"
        periodo_year:
          type: integer
        periodo_month:
          type: integer
        fecha_pedido:
          type: string
          format: date
        estado:
          type: string
          enum: [pendiente, aprobado, rechazado, cancelado]
          default: pendiente
        cantidad_items:
          type: integer
        total_bultos:
          type: integer
        total_unidades:
          type: integer
        motivo_rechazo:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        resuelto_at:
          type: string
          format: date-time
          nullable: true
        resuelto_by_usuario:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/PedidoMercaderiaLinea'
    PaginatedPedidosMercaderia:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PedidoMercaderiaItem'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
...
    PedidoMercaderiaUpsertRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [articulo_id]
            description: Debe indicar al menos un bulto o una unidad; ambos valores pueden combinarse.
            anyOf:
              - required: [cantidad_bultos]
              - required: [cantidad_unidades]
            properties:
              articulo_id:
                type: integer
              cantidad_bultos:
                type: integer
                minimum: 0
                default: 0
              cantidad_unidades:
                type: integer
                minimum: 0
                default: 0
                description: Unidades sueltas adicionales a los bultos completos.

    # --- Horario ---
...
    LegajoAdjuntoItem:
      type: object
      properties:
        id:
          type: integer
        evento_id:
          type: integer
        nombre_original:
          type: string
        mime_type:
          type: string
          example: application/pdf
        extension:
          type: string
          nullable: true
        tamano_bytes:
          type: integer
        estado:
          type: string
          enum: [activo, eliminado]
        created_at:
          type: string
          format: date-time
          nullable: true
        download_url:
          type: string
          example: /api/v1/mobile/me/legajo/adjuntos/99
    LegajoEventoItem:
      type: object
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
          nullable: true
        empleado_id:
          type: integer
          nullable: true
        tipo_id:
          type: integer
        tipo_codigo:
          type: string
        tipo_nombre:
          type: string
        fecha_evento:
          type: string
          format: date
          nullable: true
        fecha_desde:
          type: string
          format: date
          nullable: true
        fecha_hasta:
          type: string
          format: date
          nullable: true
        titulo:
          type: string
        descripcion:
          type: string
        estado:
          type: string
          enum: [vigente, anulado]
        severidad:
          type: string
          nullable: true
          enum: [grave, media, leve]
        justificacion_id:
          type: integer
          nullable: true
        adjuntos_count:
          type: integer
        adjuntos:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/LegajoAdjuntoItem'
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
    LegajoTipoEventoItem:
      type: object
      properties:
        id:
          type: integer
        codigo:
          type: string
        nombre:
          type: string
        requiere_rango_fechas:
          type: boolean
        permite_adjuntos:
          type: boolean
        habilitado_mobile:
          type: boolean
        activo:
          type: boolean
    LegajoTiposEventoResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/LegajoTipoEventoItem'
        total:
          type: integer
...
    PaginatedLegajoEventos:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/LegajoEventoItem'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        pagination:
          type: object
          properties:
            page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer
            has_more:
              type: boolean

    # --- Seguridad ---
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string

    # --- Feedback de calle ---
...
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /me/pedidos-mercaderia:
    get:
      tags: [Pedidos mercaderia]
      summary: Historial paginado de pedidos de mercaderia del empleado
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendiente, aprobado, rechazado, cancelado]
      responses:
        '200':
          description: Historial paginado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPedidosMercaderia'
        '400':
          description: Parametro de estado invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Pedidos mercaderia]
      summary: Crea el pedido del mes actual de mercaderia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoMercaderiaUpsertRequest'
      responses:
        '201':
          description: Pedido creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoMercaderiaItem'
        '400':
          description: Validacion fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Ya existe un pedido en el mes actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/pedidos-mercaderia/{id}:
    get:
      tags: [Pedidos mercaderia]
      summary: Detalle de un pedido propio de mercaderia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalle del pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoMercaderiaItem'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Pedidos mercaderia]
      summary: Actualiza los items de un pedido pendiente propio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoMercaderiaUpsertRequest'
      responses:
        '200':
          description: Pedido actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoMercaderiaItem'
        '400':
          description: Validacion fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Pedidos mercaderia]
      summary: Cancela un pedido pendiente propio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedido cancelado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoMercaderiaItem'
        '400':
          description: Validacion fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---- Horarios ----
...
  /me/legajo/tipos-evento:
    get:
      tags: [Legajo]
      summary: Tipos de evento activos para legajo
      responses:
        '200':
          description: Tipos de evento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegajoTiposEventoResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkErrorResponse'
  /me/legajo/eventos:
    get:
      tags: [Legajo]
      summary: Lista paginada de eventos del legajo
      parameters:
        - name: tipo_id
          in: query
          schema:
            type: integer
        - name: estado
          in: query
          schema:
            type: string
            enum: [vigente, anulado, all]
        - name: severidad
          in: query
          schema:
            type: string
            enum: [leve, media, grave, all]
        - name: desde
          in: query
          schema:
            type: string
            format: date
        - name: hasta
          in: query
          schema:
            type: string
            format: date
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Lista de eventos del legajo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLegajoEventos'
        '400':
          description: Parametros invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkErrorResponse'
  /me/legajo/eventos/{id}:
    get:
      tags: [Legajo]
      summary: Detalle de un evento del legajo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Evento
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/LegajoEventoItem'
                  - type: object
                    properties:
                      ok:
                        type: boolean
                        example: true
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkErrorResponse'
...
  /me/legajo/eventos-admin/tipos:
    get:
      tags: [Legajo]
      summary: Tipos habilitados para carga mobile de eventos de legajo
      description: Devuelve solo tipos activos con `habilitado_mobile = 1`.
      responses:
        '200':
          description: Tipos habilitados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegajoTiposEventoResponse'
        '403':
          description: Sin permiso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkErrorResponse'
...